# SwaggerDocForRestApi.yaml
openapi: 3.1.4
info:
  title: Swagger Reactive-Rest-Api
  description: The Open API Specification for Test services
  license:
    name: "The name of the license ..."
    url: "https:// ..."
  version: 0.0.1
servers:
  - url: https://localhost.net/
    description: Production server
  - url: https://test.localhost.net/
    description: Server for testing purposes
paths:
  /client:
    post:
      tags:
        - Client Service
      summary: Create new client
      description: Operation create new client, with status active
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
    get:
      tags:
        - Client Service
      summary: Get client list
      description: Operation return client list
      operationId: getClientList
      responses:
        '200':
          description: OK
          headers:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
  /client/{id}:
    get:
      tags:
        - Client Service
      summary: Get client by uuid
      description: Operation return client
      operationId: getClient
      responses:
        '200':
          description: OK
          headers:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    delete:
      tags:
        - Client Service
      summary: Remove client by uuid
      description: Operation change client status to closed, physical from database nothing deleted
      operationId: removeClient
      responses:
        '202':
          description: Accepted
          headers:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
  /account:
    post:
      tags:
        - Account Service
      summary: Create new client account
      description: Operation create new client account
      operationId: createNewClientAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /account/client/{id}:
    get:
      tags:
        - Account Service
      summary: Get client account list
      description: Operation return client account list by client guid
      operationId: getClientAccountList
      responses:
        '200':
          description: OK
          headers:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
  /transaction:
    post:
      tags:
        - Transaction Service
      summary: Create new transaction
      description: Operation create new debit / credit transaction. A debit entry in an account represents a transfer of value to that account, and a credit entry represents a transfer from the account.
      operationId: createTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /transaction/account/{id}:
    parameters:
      - name: page
        in: path
        description: Page number
        required: true
        example: "1"
    get:
      tags:
        - Transaction Service
      summary: Get transaction list for account
      description: Operation return paginated transaction list for selected account
      operationId: getTransactionsForAccount
      responses:
        '200':
          description: OK
          headers:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
components:
  schemas:
    ClientRequest:
      type: object
      properties:
        name:
          type: string
    Client:
      type: object
      properties:
        guid:
          type: string
        name:
          type: string
        status:
          enum:
            - "Active"
            - "Closed"
    AccountRequest:
      type: object
      properties:
        clientGuid:
          type: string
        name:
          type: string
        currency:
          type: string
    Account:
      type: object
      properties:
        guid:
          type: string
        clientGuid:
          type: string
        clientName:
          type: string
        name:
          type: string
        currency:
          type: string
    TransactionRequest:
      type: object
      properties:
        accGuid:
          type: string
        corrAccGuid:
          type: string
        trxAmount:
          type: string
        trxCurrency:
          type: string
        trxType:
          enum:
            - "Debit"
            - "Credit"
    Transaction:
      type: object
      properties:
        guid:
          type: string
        accGuid:
          type: string
        accCurrency:
          type: string
        beginAmount:
          type: string
        trxAmount:
          type: string
        trxCurrency:
          type: string
        endAmount:
          type: string
        trxType:
          enum:
            - "Debit"
            - "Credit"